+++ templates/app/template/__tests__/unit/widgets/App.js
@@ -1,7 +1,0 @@
-import App from '../../../src/widgets/App.sht';
-import renderer from 'sham-ui-test-helpers';
-
-it( 'renders correctly', () => {
-    const meta = renderer( App, {} );
-    expect( meta.toJSON() ).toMatchSnapshot();
-} );