+++ templates/app/template/__tests__/unit/widgets/Header.js
@@ -1,7 +1,0 @@
-import Header from '../../../src/widgets/Header.sht';
-import renderer from 'sham-ui-test-helpers';
-
-it( 'renders correctly', () => {
-    const meta = renderer( Header, {} );
-    expect( meta.toJSON() ).toMatchSnapshot();
-} );