+++ templates/app/template/webpack.config.js
@@ -57,14 +57,14 @@ function createConfig( isProd ) {
                 test: /\.sht$/,
                 loader: 'sham-ui-templates-loader?hot'
             }, {
-                test: /\.sfw$/,
+                test: /\.sfc$/,
                 use: [
                     { loader: 'babel-loader' },
                     {
                         loader: 'sham-ui-templates-loader?hot',
                         options: {
                             asModule: false,
-                            asSingleFileWidget: true
+                            asSingleFileComponent: true
                         }
                     }
                 ]