+++ templates/component/template/demo-app/src/initializers/main.js
@@ -1,0 +1,8 @@
+import App from '../components/App.sht';
+
+export default function() {
+    new App( {
+        ID: 'app',
+        containerSelector: 'body'
+    } );
+}