+++ templates/component/template/webpack.config.js
@@ -1,0 +1,45 @@
+const webpack = require( 'webpack' );
+
+const plugins = [
+    new webpack.NoEmitOnErrorsPlugin()
+];
+
+module.exports = {
+    entry: {
+        index: './src/{{name}}.js',
+    },
+    output: {
+        path: __dirname,
+        filename: '[name].js',
+        publicPath: '/',
+        library: ['{{name}}', '{{name}}/[name]' ],
+        libraryTarget: 'umd'
+    },
+    externals: [
+        'sham-ui'
+    ],
+    plugins: plugins,
+    module: {
+        rules: [ {
+            test: /(\.js)$/,
+            loader: 'babel-loader',
+            exclude: /(node_modules)/,
+            include: __dirname
+        }, {
+            test: /\.sht/,
+            loader: 'sham-ui-templates-loader?{}'
+        }, {
+            test: /\.sfc$/,
+            use: [
+                { loader: 'babel-loader' },
+                {
+                    loader: 'sham-ui-templates-loader?hot',
+                    options: {
+                        asModule: false,
+                        asSingleFileComponent: true
+                    }
+                }
+            ]
+        } ]
+    }
+};