+++ templates/widget/template/demo-app/src/controllers/main.js
@@ -1,8 +1,0 @@
-import App from '../widgets/App.sht';
-
-export default function() {
-    new App( {
-        ID: 'app',
-        containerSelector: 'body'
-    } );
-}